+++ s4e-web/src/app/views/map-view/sentinel-search/search-result-modal/search-result-modal.model.ts
@@ -1,0 +1,2 @@
+export const SENTINEL_SEARCH_RESULT_MODAL_ID = 'search-result-modal-id';
+